fix: update location of sha256 file in dump backup script
The script for generating database backups has been updated. It changes where the sha256 checksum file is stored. Previously, it was saved in a subdirectory with the date. Now, it is kept in the same directory as the tar.gz backup file. This provides easier access and a cleaner folder structure.